Refactor the Deserialization API into an Assertion API
