+++ betty/serde/dump.py
@@ -6,18 +6,7 @@ from __future__ import annotations
 
 from typing import TypeVar, Sequence, Mapping, overload, Literal, TypeAlias, Any
 
-
-class Void:
-    """
-    A sentinel that describes the absence of a value.
-
-    Using this sentinel allows for actual values to be ``None``. Like ``None``,
-    ``Void`` is only ever used through its type, and never instantiated.
-    """
-
-    def __new__(cls):  # pragma: no cover  # noqa D102
-        raise RuntimeError("The Void sentinel cannot be instantiated.")
-
+from betty.typing import Void
 
 #: The Python types that define a serialized dump.
 DumpType: TypeAlias = bool | int | float | str | None | list["Dump"] | dict[str, "Dump"]
